// frontend/src/types/CartItem.ts
export interface CartItem {
  bookID: number;
  title: string;
  price: number;
  quantity: number;
}

// bookID: number;
// title: string;
// author: string;
// publisher: string;
// isbn: string;
// classification: string;
// category: string;
// pageCount: number;
// price: number;